packages/kubebox/pyproject.toml
[tool.poetry]
name = "kubebox"
version = "0.0.3"
description = "Open source k8s sandbox for LLM agents"
authors = ["Lucas Jaggernauth <lukejagg@gmail.com>"]
readme = "README.md"
repository = "https://github.com/lukejagg/kubebox"
classifiers = [
    "Typing :: Typed",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/lukejagg/kubebox"
Repository = "https://github.com/lukejagg/kubebox"

[tool.poetry.dependencies]
python = "^3.8"
kubernetes = "^30.0.0"
fastapi-socketio = "^0.0.10"
aiohttp = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
